requirements_gpu.txt
tensorflow-gpu
keras
numpy
matplotlib
scipy
cv2
